implement the MVC structure to the project
